FbAnalizingSite/requirements.txt
asgiref==3.2.7
astroid==2.4.2
beautifulsoup4==4.9.1
cachetools==4.1.0
certifi==2020.4.5.2
chardet==3.0.4
Django==3.0.7
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
google-api-core==1.20.0
google-auth==1.16.1
google-cloud-language==1.3.0
googleapis-common-protos==1.52.0
grpcio==1.29.0
html5lib==1.0.1
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pylint==2.5.3
pytz==2020.1
requests==2.23.0
rsa==4.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.9
webencodings==0.5.1
wrapt==1.12.1
